# types: improve the performance of sprintName by ~30% and halve allocs