This PR improves performance and memory usage by lazily allocating the `strings.Builder` and by inlining the `writeDomainNameByte()` logic.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkSprintName-12              180           120           -33.33%
BenchmarkSprintName_NoEscape-12     198           82.4          -58.38%

benchmark                           old allocs     new allocs     delta
BenchmarkSprintName-12              2              1              -50.00%
BenchmarkSprintName_NoEscape-12     1              0              -100.00%

benchmark                           old bytes     new bytes     delta
BenchmarkSprintName-12              48            32            -33.33%
BenchmarkSprintName_NoEscape-12     32            0             -100.00%
```

**Edit:** forgot to include the results of `BenchmarkSprintName_NoEscape`.