@tmthrgd 2019-09-21
Can you use [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) to compare ~10 benchmark before and after runs rather than benchcmp please? (benchstat is more statistically sound and accounts for noise).