@charlievieth 2019-09-21
> Can you use [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) to compare ~10 benchmark before and after runs rather than benchcmp please? (benchstat is more statistically sound and accounts for noise).

Interesting, never heard of `benchstat`.  I highly doubt it will make a difference here since the gain comes from the reduction in allocs, which isn't variable or subject noise, but I'll run it out of curiosity.