@charlievieth 2019-09-21
Ran `benchstat` as predicted it didn't really show anything - cool tool nonetheless so thanks for cluing me in about it.

PR [`cev/sprintname-perf`](https://github.com/charlievieth/dns/tree/cev/sprintname-perf)
```
name                    time/op
SprintName-12            122ns ± 3%
SprintName_NoEscape-12  84.2ns ± 3%

name                    alloc/op
SprintName-12            32.0B ± 0%
SprintName_NoEscape-12   0.00B

name                    allocs/op
SprintName-12             1.00 ± 0%
SprintName_NoEscape-12    0.00
```

master@ 2acbc9eff343291e671c1a9ba7be3201bf66dfd6
```
name                    time/op
SprintName-12           185ns ± 1%
SprintName_NoEscape-12  202ns ± 1%

name                    alloc/op
SprintName-12           48.0B ± 0%
SprintName_NoEscape-12  32.0B ± 0%

name                    allocs/op
SprintName-12            2.00 ± 0%
SprintName_NoEscape-12   1.00 ± 0%
```